Cursor editing

+++ src/terminal/cursor.rs
@@ -3,11 +3,11 @@ use termwiz::escape::csi::CursorStyle;
 /// Cursor object to store cursor position and style
 /// Allows for storing, restoring etc with positions as well
 pub struct TerminalCursor {
-    pub x: u32,
-    pub y: u32,
+    pub x: usize,
+    pub y: usize,
 
-    pub saved_x: u32,
-    pub saved_y: u32,
+    pub saved_x: usize,
+    pub saved_y: usize,
 
     pub style: CursorStyle,
 }
@@ -25,36 +25,36 @@ impl TerminalCursor {
 
     /// Sets the cursor position
     pub fn set(&mut self, x: u32, y: u32) {
-        self.x = x;
-        self.y = y;
+        self.x = x as usize;
+        self.y = y as usize;
     }
 
     pub fn set_x(&mut self, x: u32) {
-        self.x = x
+        self.x = x as usize
     }
 
     pub fn set_y(&mut self, y: u32) {
-        self.y = y
+        self.y = y as usize
     }
 
     /// Shifts the cursor down
     pub fn shift_down(&mut self, amount: u32) {
-        self.y += amount;
+        self.y += amount as usize;
     }
 
     /// Shifts the cursor right
     pub fn shift_right(&mut self, amount: u32) {
-        self.x += amount
+        self.x += amount as usize
     }
 
     /// Shifts the cursor up
     pub fn shift_up(&mut self, amount: u32) {
-        self.y = self.y.checked_sub(amount).unwrap();
+        self.y = self.y.checked_sub(amount as usize).unwrap();
     }
 
     /// Shifts the cursor left
     pub fn shift_left(&mut self, amount: u32) {
-        self.x = self.x.checked_sub(amount).unwrap();
+        self.x = self.x.checked_sub(amount as usize).unwrap();
     }
 
     pub fn set_style(&mut self, style: CursorStyle) {
